feature: Standardize DebugLogAction for AbstractAction
