+++ src/index.ts
@@ -31,6 +31,7 @@ export * from './validations/message-validators/follow/FollowValidators';
  * Actions
  */
 export * from './actions/AbstractAction';
+export * from './actions/LoggingActions';
 export * from './actions/standard-bsky-actions/SkeetActions';
 export * from './actions/standard-bsky-actions/LikeActions';
 export * from './actions/standard-bsky-actions/ReskeetActions';
@@ -38,7 +39,7 @@ export * from './actions/standard-bsky-actions/FollowActions';
 export * from './actions/message-actions/AbstractMessageAction';
 export * from './actions/message-actions/TestAction';
 export * from './actions/message-actions/FunctionAction';
-export * from './actions/message-actions/LoggingActions';
+export * from './actions/message-actions/MessageLoggingActions';
 export * from './actions/message-actions/post/SkeetMessageActions';
 
 /**
